+++ .devcontainer/features/test/git-tools/test.sh
@@ -8,5 +8,6 @@ check "gh version" gh version
 check "git-crypt version" git-crypt --version
 check "pre-commit version" pre-commit --version
 check "detect-secrets version" detect-secrets --version
+check "gitmoji version" gitmoji --version
 
 reportResults