+++ containers/daap-presigned-url/src/var/task/presigned_url.py
@@ -2,6 +2,7 @@ import json
 import os
 import uuid
 from datetime import datetime
+from pathlib import PurePath
 
 import boto3
 from data_platform_logging import DataPlatformLogger
@@ -20,30 +21,46 @@ logger = DataPlatformLogger(
 def handler(event, context):
     data_product_name = event["pathParameters"].get("data-product-name")
     table_name = event["pathParameters"].get("table-name")
-    md5 = str(event["body"].get("contentMD5"))
+    body = json.loads(event.get("body"))
+    md5 = str(body.get("contentMD5"))
+    filename = body.get("filename")
     amz_date = datetime.utcnow()
     formatted_date = amz_date.strftime("%Y%m%dT%H%M%SZ")
     uuid_value = uuid.uuid4()
 
-    if not isinstance(data_product_name, str) or not isinstance(table_name, str):
+    if (
+        not isinstance(data_product_name, str)
+        or not isinstance(table_name, str)
+        or not isinstance(filename, str)
+    ):
         return {
             "statusCode": 400,
             "headers": {"Content-Type": "application/json"},
             "body": json.dumps(
                 {
                     "error": {
-                        "message": "data product name or table name value are not"
+                        "message": "data product name, table name or filename are not"
                         + " convertible to string type."
                     }
                 }
             ),
         }
 
+    file_extension = PurePath(filename).suffix
+    if file_extension == "":
+        return {
+            "statusCode": 400,
+            "headers": {"Content-Type": "application/json"},
+            "body": json.dumps({"error": {"message": "file extension is invalid."}}),
+        }
+
     element = DataProductElement.load(
         data_product_name=data_product_name, element_name=table_name
     )
     data_product = element.data_product
-    raw_data_path = element.raw_data_path(timestamp=amz_date, uuid_value=uuid_value)
+    raw_data_path = element.raw_data_path(
+        timestamp=amz_date, uuid_value=uuid_value, file_extension=file_extension
+    )
 
     fields = {
         "x-amz-server-side-encryption": "AES256",